python_diplom/requirements.txt
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
PyYAML==6.0
requests==2.28.1
ujson==5.5.0
urllib3==1.26.12


